+++ corehq/motech/repeaters/tests/test_refer_case_repeater.py
@@ -1,61 +1,15 @@
 import uuid
-from http.client import responses
 from unittest.mock import patch, Mock
 from xml.etree import cElementTree as ElementTree
 
-from django.test import SimpleTestCase
 from testil import eq
 
 from casexml.apps.case.mock import CaseBlock
 from casexml.apps.case.xml import V2_NAMESPACE
 from corehq.form_processor.interfaces.dbaccessors import CaseAccessors
 from corehq.form_processor.models import CommCareCaseSQL
-from corehq.motech.repeater_helpers import RepeaterResponse
 from corehq.motech.repeaters.models import ReferCaseRepeater
 from corehq.motech.repeaters.repeater_generators import ReferCasePayloadGenerator
-from corehq.util.test_utils import generate_cases
-from couchforms.openrosa_response import ResponseNature
-
-SUCCESS = f"""
-    <OpenRosaResponse xmlns="http://openrosa.org/http/response">
-        <message nature="{ResponseNature.SUBMIT_SUCCESS}">success</message>
-    </OpenRosaResponse>
-    """
-
-V3_RETRY_ERROR = f"""
-    <OpenRosaResponse xmlns="http://openrosa.org/http/response">
-        <message nature="{ResponseNature.POST_PROCESSING_FAILURE}">success</message>
-    </OpenRosaResponse>
-    """
-
-V3_ERROR = f"""
-    <OpenRosaResponse xmlns="http://openrosa.org/http/response">
-        <message nature="{ResponseNature.PROCESSING_FAILURE}">success</message>
-    </OpenRosaResponse>
-    """
-
-V2_ERROR = f"""
-    <OpenRosaResponse xmlns="http://openrosa.org/http/response">
-        <message nature="{ResponseNature.SUBMIT_ERROR}">success</message>
-    </OpenRosaResponse>
-    """
-
-
-class TestReferCaseRepeater(SimpleTestCase):
-    pass
-
-
-@generate_cases([
-    (201, SUCCESS, RepeaterResponse(201, 'Created', '')),
-    (201, V2_ERROR, RepeaterResponse(422, ResponseNature.SUBMIT_ERROR, '', False)),
-    (422, V3_ERROR, RepeaterResponse(422, ResponseNature.PROCESSING_FAILURE, '', False)),
-    (422, V3_RETRY_ERROR, RepeaterResponse(422, ResponseNature.POST_PROCESSING_FAILURE, '', True)),
-], TestReferCaseRepeater)
-def test_get_response(self, status_code, text, expected_response):
-    response = ReferCaseRepeater.get_response(RepeaterResponse(status_code, responses[status_code], text))
-    self.assertEqual(response.status_code, expected_response.status_code)
-    self.assertEqual(response.reason, expected_response.reason)
-    self.assertEqual(response.retry, expected_response.retry)
 
 
 def test_refer_case_payload_generator_no_previous_transfer():
